refactor: add blobs.Store abstraction
